create a landing page for the questioner app
Landing page for the Questioner app. The page also contains fields for user sign in and sign up.

The standalone sign in and sign up pages are active

Changes were made to the folder structure to accomodate the new page.

Also, minor design changes were made to the comment button.

Pivotal Tracker: https://www.pivotaltracker.com/story/show/163032362
