+++ admin/index.html
@@ -6,8 +6,8 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <link href="https://fonts.googleapis.com/css?family=Fira+Sans|Josefin+Sans|Slabo+27px|Zilla+Slab" rel="stylesheet">
-  <link rel="stylesheet" href="../style.css">
-  <title>Questioner</title>
+  <link rel="stylesheet" href="../home/style.css">
+  <title>Questioner | Admin</title>
 </head>
 
 <body>
@@ -142,7 +142,7 @@
       <div class="profile-container-main">
         <div id="profile-picture-container">
           <label id="profile-picture-label" for="profile-picture-input"></label>
-          <img id="profile-picture" src="./assets/profile.svg" alt="profile pic" srcset="">
+          <img id="profile-picture" src="../assets/profile.svg" alt="profile pic" srcset="">
           <input id="profile-picture-input" accept="image/jpeg, image/png" type="file">
         </div>
       </div>
