+++ index.html
@@ -4,95 +4,86 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <link href="https://fonts.googleapis.com/css?family=Fira+Sans|Josefin+Sans|Slabo+27px|Zilla+Slab" rel="stylesheet">
   <link rel="stylesheet" href="style.css">
-  <title>Questioner</title>
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <title>Questioner | Welcome</title>
 </head>
 
 <body>
-  <div class="content-container">
-    <header class="header">
-      <p class="logo">Questioner</p>
-    </header>
-    <nav>
-      <span class="tab-selector tab-active">
-        <img class="icon-img" src="./assets/meetup.svg" alt="" srcset="">
-        <span class="tab-selector-text">meets</span>
-      </span>
-      <span class="tab-selector">
-        <img class="icon-img" src="./assets/booked.svg" alt="" srcset="">
-        <span class="tab-selector-text">booked</span>
-      </span>
-      <span class="tab-selector">
-        <img class="icon-img" src="./assets/notifications.svg" alt="" srcset="">
-        <span class="tab-selector-text">notifications</span>
-      </span>
-      <span class="tab-selector" id="tab-selector-profile">
-        <img class="icon-img profile-icon-img" id="profile-icon" src="./assets/profile.svg" alt="" srcset="">
-        <span class="tab-selector-text">profile</span>
-      </span>
-    </nav>
-
-    <section class="main-section section-showing" id="meets-section">
-    </section>
-
-    <section class="main-section" id="booked-section">
-    </section>
-
-    <section class="main-section" id="notif-section">
-    </section>
-
-    <section class="main-section" id="profile-section">
-      <button class="profile-button" type="button">logout</button>
-      <div class="profile-container-main">
-        <div id="profile-picture-container">
-          <label id="profile-picture-label" for="profile-picture-input"></label>
-          <img id="profile-picture" src="./assets/profile.svg" alt="profile pic" srcset="">
-          <input id="profile-picture-input" accept="image/jpeg, image/png" type="file">
+  <header class="main header">
+    <p>Questioner</p>
+  </header>
+  <div class="text-container">
+    <p>Crowd-source Questions</p>
+    <p>Find meetups, ask questions, comment, vote, RSVP</p>
+    <P>Ask a question now!</P>
+  </div>
+  <div class="image container"></div>
+  <div class="main-sign-container">
+    <div class="tab-container">
+      <div class="tab-signup tab-item active">
+        <p>Sign Up</p>
+      </div>
+      <div class="tab-signin tab-item">
+        <p>Sign In</p>
+      </div>
+    </div>
+    <div class="sign-container signup">
+      <div class="signup-container">
+        <div class="input-container">
+          <label for="input-username" class="input-label">Username</label>
+          <input type="text" id="input-username" pointer="username" class="form-input">
+        </div>
+        <div class="input-container">
+          <label for="input-firstname" class="input-label">First name</label>
+          <input type="text" id="input-firstname" pointer="firstname" class="form-input">
+        </div>
+        <div class="input-container">
+          <label for="input-secondname" class="input-label">Second name</label>
+          <input type="text" id="input-secondname" pointer="secondname" class="form-input">
+        </div>
+        <div class="input-container">
+          <label for="input-othername" class="input-label">Other name</label>
+          <input type="text" id="input-othername" pointer="othername" class="form-input">
+        </div>
+        <div class="input-container">
+          <label for="input-email" class="input-label">Email</label>
+          <input type="text" id="input-email" pointer="email" class="form-input">
+        </div>
+        <div class="input-container">
+          <label for="input-phoneNumber" class="input-label">Phone</label>
+          <input type="text" id="input-phoneNumber" pointer="phoneNumber" class="form-input">
+        </div>
+        <div class="input-container">
+          <label for="input-password1" class="input-label">Password</label>
+          <input type="password" id="input-password1" pointer="password" class="form-input">
+        </div>
+        <div class="input-container">
+          <label for="input-password2" class="input-label">Confirm password</label>
+          <input type="password" id="input-password2" pointer="password" class="form-input">
+        </div>
+        <div class="button-container">
+          <button class="submit-button" id="signup-button" type="button">Join</button>
         </div>
       </div>
-      <div class="profile-container">
-        <label for="profile-username" class="profile-input-label">Username:</label>
-        <input pointer="username" type="text" id="profile-username" class="profile-input" disabled>
-        <button input="profile-username" class="profile-edit-button" type="button">edit</button>
+      <div class="signin-container">
+        <div class="signin-container-2">
+        <div class="input-container">
+          <label for="input-email1" class="input-label">Email</label>
+          <input type="text" id="input-email1" pointer="email" class="form-input">
+        </div>
+        <div class="input-container">
+          <label for="input-password3" class="input-label">Password</label>
+          <input type="password" id="input-password3" pointer="password" class="form-input">
+        </div>
+        <div class="button-container">
+        <button class="submit-button" type="button">Enter</button>
+        </div>
+        </div>
       </div>
-      <div class="profile-container">
-        <label for="profile-first-name" class="profile-input-label">First name:</label>
-        <input pointer="firstname" type="text" id="profile-first-name" class="profile-input" disabled>
-        <button input="profile-first-name" class="profile-edit-button" type="button">edit</button>
-      </div>
-      <div class="profile-container">
-        <label for="profile-second-name" class="profile-input-label">Second name:</label>
-        <input pointer="secondname" type="text" id="profile-second-name" class="profile-input" disabled>
-        <button input="profile-second-name" class="profile-edit-button" type="button">edit</button>
-      </div>
-      <div class="profile-container">
-        <label for="profile-third-name" class="profile-input-label">Other names:</label>
-        <input pointer="othername" type="text" id="profile-third-name" class="profile-input" disabled>
-        <button input="profile-third-name" class="profile-edit-button" type="button">edit</button>
-      </div>
-      <div class="profile-container">
-        <label for="profile-email" class="profile-input-label">Email:</label>
-        <input pointer="email" type="text" id="profile-email" class="profile-input" disabled>
-        <button input="profile-email" class="profile-edit-button" type="button">edit</button>
-      </div>
-      <div class="profile-container">
-        <label for="profile-phone" class="profile-input-label">Phone:</label>
-        <input pointer="phoneNumber" type="text" id="profile-phone" class="profile-input" disabled>
-        <button input="profile-phone" class="profile-edit-button" type="button">edit</button>
-      </div>
-    </section>
-    <section class="meet-expanded-section main-section" id="meet-expanded">
-      <div id="meet-expanded-container">
-      </div>
-    </section>
-
+    </div>
   </div>
-
-  <footer>
-    <p>questioner</p>
-  </footer>
   <script type="module" src="script.js"></script>
 </body>
